processing/requirements.txt
connexion[flask,uvicorn,swagger-ui]
PyYAML
requests
httpx
APScheduler